Merged PR 858: native-platform-installer
Tests and example-code no longer depend directly on any version of the native-platform.
Instead, they delegate the task of installing and loading the addon to the new tools/native-platform-installer
package. They call NativePlatformRegistry.loadAndRegisterStandardNativePlatformFromTools method
to load the native platform.
As a result of this change, we no longer need the "aggregator" packages, such as imodeljs-native-platform-node.
They will be removed on the hg side.