Merged PR 8713: UI: Display arrays & structs in property grid
This makes property grid show structs and arrays instead of [Object object]. The result is not very user friendly, but still much better than nothing until full support for arrays and struct is implemented. Leaving for @<Josh Schifter> to decide whether we want to merge this into master before YII or not.