Merged PR 9618: Update to npm 5.8.0
There is an issue with npm 5.6.0 where linux and windows would produce inconsistent sha512 hashes for the integrity of internal packages. I think this has to do with the Rush/npm interop but Rush is not going to fix it since npm@5 is not going to be supported long-term. It appears that 5.8 does not have the regression.