Merged PR 9772: Speedup run time of UI and Presentation unit tests.
This should significantly speed up the _execution_ of the UI and presentation tests (especially ui-ninezone).  I'm still looking into ways we can speedup the _build_ time.

The main culprit here seemed to be using sourcemaps to determine the snapshot file location.
This is just too expensive to do before _every_ test (and suite), so for now I've just replaced it with a simpler pair of string replaces.
This seems to work just as well - I'm not seeing any misplaced snapshot files - and it's _**significantly**_ faster.

I also took this opportunity to remove all unused snapshots.