Merged PR 23238: View Attributes fixes (2d view exception & updated issues)
When opening Display Test App's View Attributes menu in a 2d view, exception would be thrown, crashing the app.  Also, the skybox attributes would fail to update correctly when a iModel was opened that did not have the skybox enabled, or an iModel was opened while the menu was open.
Also fixed issue with hiding flags on the View Attribute menu.

Related work items: #102986, #102987