Merged PR 24724: UI: Fix indefinite Tree component rendering (attempt #2)
My second attempt on this... I think the problem is that updating checkboxes from getDerivedStateFromProps causes a setState to be called which resets state changes made in getDerivedStateFromProps. And this causes an infinite loop. On a local debug build I'm not seeing this because the `updateTreeCheckboxes` call is handled asynchronously, and in production build it seems to be handled synchronously. Still no idea why there's a difference..