Merged PR 33005: Remove need to cache tool setting properties
Remove need to cache tool setting properties. By not triggering the active tool changed callback until after the Primitive, Viewing, of Collection tools has been set in the ToolAdmin the UI code is be able to retrieve the active tool and then see if it provides tool setting properties. Brien confirmed that it was OK to delay the event firing until the active tool was set.

Related work items: #137311