Merged PR 37636: Clean up @deprecated messages for lint rule
This PR attempts to set a convention for handling the deprecation message as there are competing concerns:

1. The deprecation message is used verbatim by the deprecation lint rule so extra formatting like `[]($backend)` makes the lint message look awkward.
2. The deprecation message is also used by the SDK docs, so linking to the replacement method would be preferred.

The current proposal is to only use "double-square-bracket linking" (for example: `[[replacement]]`) to have a documentation link and to have something that looks reasonable in the lint message. Other special SDK doc formatting should be avoided. The guideline docs have been updated accordingly in this PR.

This PR also addresses a few lint warnings.