Merged PR 48794: Added Table column filtering support
The column filter row in the react-data-grid component is used. The react-data-grid-addons package provides Text, Numeric, SingleSelect and MultiSelect filter renderers. The MultiSelect renderer leaves much to be desired, so I will develop a custom renderer for multi-select using checkboxes.

Much of the column filtering code was brought over from DgnClientFx and cleaned up for iModel.js.

There are 3 new methods in TableDataProvider for column filtering:
* applyFilterDescriptors
* getDistinctValues
* getPropertyDisplayValueExpression

The ECExpression generation support was left in. If it is not needed, we should remove it.

Related work items: #28809