Merged PR 56981: Revert openid-client in imodeljs-clients-backend
Reverted the upgrade only for @bentley/imodeljs-clients-backend so it continues to work for previous Electron versions.  When iModel.js 2.0 comes out and officially update to Electron 6.x (i.e. Node 12) this library can be moved up as well.

The oidc-signin-tool will stay at 3.x and skip Rush check to allow it.