Merged PR 60841: Fix for nine-zone sample app not displaying backstage items
Due to the component load order the nine-zone sample app was not showing proper backstage items.
This was because manager.onChanged was triggered before useEffect in useBackstageItems had set up listener for item changes. Added a dependency on manager.items so the items are properly updated when the manager items are set, even if this occurs before the onChanged handler is set.