Merged PR 72799: Add missing tildes in sass imports
The change in @bentley/ui-components seems to have been missed when others were updated. It would be needed for consumers who are using the forked react-scripts.

The test apps may not need the changes if they are not using the new react-scripts, but I updated all of them for consistency