Merged PR 76111: findByKey instead of findByToken
findByKey instead of findByToken

Another change is to make the `key` member of `IModelToken` and `IModelTokenProps` required. I don't really see the point of the token if the key is not defined. I also didn't see the point of converting from `IModelTokenProps` to `IModelToken` only to use the key. Hopefully these changes simplify things.

> Note: I have another refactor in progress to minimize the use of `IModelToken`, but I wanted to get this part approved first.