Make IModelHost.startup() and IModelHost.shutdown async.
Make `IModelHost.startup()` and `IModelHost.shutdown()` async.

Today, `IModelHost.shutdown()` does not actually ensure that all timers and async tasks started by a backend (such as tile uploads) are completed (and/or stopped).  While we may not have time to address all of this before 2.0, we should at least change the signatures of `IModelHost.startup()` and `IModelHost.shutdown()` now so we have the flexibility of fixing this in a future 2.x release without introducing another breaking change.  This will also give us a lot more flexibility to add other async initialization/termination logic.

Note that this is a **breaking public API change for 2.0**.

See !81165  for similar (frontend) changes to `IModelApp`.