Merged PR 82407: Addon 2.0.26
- Presentation: Reacting to rulesets' modification

Add ability to modify rulesets on the frontend and get a list of modifications that need to be made in components to represent the changed rules.

Related imodel02 PR: !81299
- VSTS #301087 - Reaction to native addon taking the ChangeSetKind/ChangesType instead of a boolean.
- Merge branch 'master' into imodel02
- Presentation: Enable full stack tests for update on rulesets modification

Related imodel02 PR: !81852
- rework BriefcaseId scheme

undefined
- Merge from imodel02
- Merged PR 81367: VSTS #301087 - Reaction to native addon taking the ChangeSetKind/ChangesType

VSTS #301087 - Reaction to native addon taking the ChangeSetKind/ChangesType instead of a boolean.

Related work items: #301087
- Merge commit 'a28722ecc3f74ee7c2217ce4a04b79aba4f22886' into addon_2_0_26
- addon 2.0.26

Related work items: #28780, #301087