Merged PR 95426: Update to latest versions of openid-client and superagent packages.
Update to latest versions of openid-client and superagent packages.

The older versions we were using were causing the following node deprecation warning in Node 12:
```
[DEP0066] DeprecationWarning: OutgoingMessage.prototype._headers is deprecated
```

Technically, in the case of openid-client, this was caused by an indirect dependency (`timed-out`, a dependency of `got`, which is a dependency of `openid-client`).  Updating to the latest `openid-client` pulls the fix though.