Merged PR 105967: Ability to fit content of panel widgets
This PR adds `preferredPanelSize` flag to `WidgetProps`. When `preferredPanelSize="fit-content"` is provided to a widget (or i.e. tool settings widget) and such widget is rendered in a panel, it will have a reduced size (based on widget content size).

### Demo:
In `ui-test-app` `ToolSettings` widget of `ViewsFrontstage` is configured with this new behavior.

Related work items: #382043