Merged PR 112848: Fix debugging launch config for new VSCode auto-attach mechanism.
The improved auto-attach in VS Code 1.49.0 broke all of our backend debugging configs.  I just found this quick and dirty fix to get everyone unblocked for now, but I'm still planning to spend the rest of the day updating this to use the new debugger settings to clean up launch.json and improve the debugging experience.