Merged PR 117218: Added a new component to handle the authentication to a WMS...
Merged PR 115872: Added a new component to handle the authentication to a WMS server

Added a new component to handle the authentication to a WMS server.
Right now, only basic is supported. This component is not displayed by default.

Change the logic of a useEffect in MapLayerManager to remove the react's error message in the consol

Cherry-picked from commit `0658e1ac`.