UI: Update VirtualizedPropertyGrid styles to accomodate iTwinUI-react components (#1957)
As a part of iTwinUI adoption, existing `ExpandableBlock`, which was used by the property grid, was deprecated and replaced with an equivalent component from `iTwinUI-react` package. This caused a number of layout issues in property grid as the new component has completely different appearance.

This change restyles the property grid to match its new expandable block component and makes it easier to reason about the property grid node layout by eliminating most of the overlap between elements.