Add User preferences API to replace settings service (#2291)
This PR makes 4 main changes;

- Drops the Product Settings Service client
- Remove the IModelApp.setting API in favor of a new IModelApp.userPreferences API that is more focused directly on the user configurability of an application versus conflating both user preferences and shared configuration.
- Remove, or update, the various wrappers in place within the Presentation and AppUI code.
- Move the MapLayersSettingsService implementation out of frontend into the Map Layers Extension along with the "sharedSetting" functionality

The overall strategy moving forward for the concept we called "Settings" before 3.0 is, make a clear distinction between what is within control of the user of the application (i.e. their own preferences) and what is within the control of the admin of an iTwin/iModel (i.e. workspace configuration).

The goal of the UserPreferences API is to make it easily adaptable by any backing implementation. The easiest of which would be to use localStorage for caching of those preferences locally as the API is modeled after the web standard Storage API.

This switch of focus did break the existing Map Layers functionality as at a core level we do not yet have a iTwin config system that can be leveraged. For this reason (and I think it's better separation), I moved the MapLayersSettingService class to the Map Layer Extension and users of the Extension will need to configure an implementation of the interface.